fixed assigning raw response data + PatronValidate
Fixed an error that would occur when trying to set the raw response data
on responses when the response was null or unable to be deserialized.
Also improved the way PatronValidate worked by creating a custom
response object with the barcode and ValidPatron values set when the
response from Polaris was null.